refactor(hooks): split use-challenge-sandbox into focused sub-hooks
Extracted three single-responsibility hooks from use-challenge-sandbox
(628 LOC) per rubber-duck-reviewed Phase 6 plan:

- use-evaluation: polling, mount recovery, job cancellation. Uses a
  generation token so resetEvaluation() ignores late polling responses
  (latent race in the prior implementation). Takes a lazy getFiles()
  accessor so the evaluate() callback identity stays stable across
  keystrokes.
- use-hints: progressive hint conversation. Takes getCurrentCode() so
  fast file switches between user actions resolve correctly.
- use-solution-generation: debug-mode AI solver.

use-challenge-sandbox becomes a thin coordinator (102 LOC) that
composes useWorkspace + the three sub-hooks and orchestrates reset.
Public return shape is preserved; consumers (ChallengeSandbox.tsx,
hint-display, ResultsPanel, evaluation-result-display) unchanged.

Baseline tightened: use-challenge-sandbox removed from baseline.
